[Task 1862943] new database-lib to extract common codes between MySQL and SQL: 1. new database-lib module. 2. move database files under common-lib database folder to database lib. 3. remove ISqlServer interface.
