+++ azure-toolkit-libs/azure-toolkit-database-lib/pom.xml
@@ -9,22 +9,13 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>azure-toolkit-sqlserver-lib</artifactId>
+    <artifactId>azure-toolkit-database-lib</artifactId>
 
     <dependencies>
-        <dependency>
-            <groupId>com.azure.resourcemanager</groupId>
-            <artifactId>azure-resourcemanager-sql</artifactId>
-        </dependency>
 
         <dependency>
             <groupId>com.microsoft.azure</groupId>
             <artifactId>azure-toolkit-common-lib</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>com.microsoft.azure</groupId>
-            <artifactId>azure-toolkit-database-lib</artifactId>
         </dependency>
 
         <dependency>
@@ -43,14 +34,6 @@
     <build>
         <plugins>
             <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-dependency-plugin</artifactId>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-source-plugin</artifactId>
-            </plugin>
-            <plugin>
                 <!-- https://mvnrepository.com/artifact/org.codehaus.mojo/aspectj-maven-plugin -->
                 <!-- http://www.quabr.com/62976155/aspectj-maven-plugin-1-11-missing-tools-jar-issue-with-jdk-11 -->
                 <groupId>com.nickwongdev</groupId>
@@ -66,12 +49,6 @@
                     <outxml>true</outxml>
                     <forceAjcCompile>true</forceAjcCompile>
                     <sources/><!-- this is important!-->
-                    <aspectLibraries>
-                        <aspectLibrary>
-                            <groupId>com.microsoft.azure</groupId>
-                            <artifactId>azure-toolkit-common-lib</artifactId>
-                        </aspectLibrary>
-                    </aspectLibraries>
                 </configuration>
                 <executions>
                     <execution>
@@ -102,8 +79,15 @@
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-dependency-plugin</artifactId>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-source-plugin</artifactId>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-javadoc-plugin</artifactId>
-                <version>2.9.1</version>
                 <configuration>
                     <show>private</show>
                     <failOnError>false</failOnError>
@@ -114,9 +98,13 @@
                         <goals>
                             <goal>jar</goal>
                         </goals>
+                        <configuration>
+                            <additionalparam>${javadoc.opts}</additionalparam>
+                        </configuration>
                     </execution>
                 </executions>
             </plugin>
         </plugins>
     </build>
+
 </project>
