+++ azure-toolkit-libs/azure-toolkit-mysql-lib/src/main/java/com/microsoft/azure/toolkit/lib/mysql/service/MySqlFirewallRules.java
@@ -8,7 +8,7 @@ package com.microsoft.azure.toolkit.lib.mysql.service;
 import com.azure.resourcemanager.mysql.MySqlManager;
 import com.azure.resourcemanager.mysql.models.FirewallRule;
 import com.google.common.base.Preconditions;
-import com.microsoft.azure.toolkit.lib.common.database.FirewallRuleEntity;
+import com.microsoft.azure.toolkit.lib.database.entity.FirewallRuleEntity;
 import com.microsoft.azure.toolkit.lib.mysql.model.MySqlEntity;
 import lombok.AllArgsConstructor;
 import org.apache.commons.lang3.StringUtils;
